ユーザーと記事の紐付けを作成途中

+++ db/migrate/20200916120514_remove_account_id_from_posts.rb
@@ -1,5 +1,5 @@
 class RemoveAccountIdFromPosts < ActiveRecord::Migration[6.0]
   def change
-    remove_column :posts, :account_id, :integer
+    remove_column :posts, :account_id, :bigint
   end
 end
